+++ BP/scripts/main.js
@@ -1,5 +1,6 @@
 import "./DoriosCore/index.js";
 import * as DoriosLib from "./DoriosLib/index.js";
+import "./config/recipes/crafter.js";
 import "./machinery/index.js";
 
 DoriosLib.registry.install();